+++ win32/wintap.c
@@ -259,19 +259,19 @@ int open_wintap(struct tuntap_dev *device,
 
 /* ************************************************ */
 
-int tuntap_read(struct tuntap_dev *tuntap, unsigned char *buf, int len) {
+ssize_t tuntap_read(struct tuntap_dev *tuntap, unsigned char *buf, size_t len) {
     DWORD read_size, last_err;
 
     ResetEvent(tuntap->overlap_read.hEvent);
-    if (ReadFile(tuntap->device_handle, buf, len, &read_size, &tuntap->overlap_read)) {
+    if (ReadFile(tuntap->device_handle, buf, (DWORD) len, &read_size, &tuntap->overlap_read)) {
         //printf("tun_read(len=%d)\n", read_size);
-        return read_size;
+        return (ssize_t) read_size;
     }
     switch (last_err = GetLastError()) {
     case ERROR_IO_PENDING:
         WaitForSingleObject(tuntap->overlap_read.hEvent, INFINITE);
         GetOverlappedResult(tuntap->device_handle, &tuntap->overlap_read, &read_size, FALSE);
-        return read_size;
+        return (ssize_t) read_size;
         break;
     default:
         printf("GetLastError() returned %d\n", last_err);
@@ -282,7 +282,7 @@ int tuntap_read(struct tuntap_dev *tuntap, unsigned char *buf, int len) {
 }
 /* ************************************************ */
 
-int tuntap_write(struct tuntap_dev *tuntap, unsigned char *buf, int len) {
+ssize_t tuntap_write(struct tuntap_dev *tuntap, unsigned char *buf, size_t len) {
     DWORD write_size;
 
     //printf("tun_write(len=%d)\n", len);
@@ -290,19 +290,19 @@ int tuntap_write(struct tuntap_dev *tuntap, unsigned char *buf, int len) {
     ResetEvent(tuntap->overlap_write.hEvent);
     if (WriteFile(tuntap->device_handle,
         buf,
-        len,
+        (DWORD) len,
         &write_size,
         &tuntap->overlap_write))
     {
         //printf("DONE tun_write(len=%d)\n", write_size);
-        return write_size;
+        return (ssize_t) write_size;
     }
 
     switch (GetLastError()) {
     case ERROR_IO_PENDING:
         WaitForSingleObject(tuntap->overlap_write.hEvent, INFINITE);
         GetOverlappedResult(tuntap->device_handle, &tuntap->overlap_write, &write_size, FALSE);
-        return write_size;
+        return (ssize_t) write_size;
         break;
     default:
         break;
